school-desk-1

+++ app/__init__.py
@@ -6,7 +6,7 @@ from flask import Flask
 from flask_restful import Api
 from flask_sqlalchemy import SQLAlchemy
 from  flask_login import LoginManager
-from app.model.Base import db
+from app.model.Base import db #db只是在views中生效
 
 
 from app.model.Students import Students
@@ -25,11 +25,11 @@ from sqlalchemy.orm import sessionmaker
 db_connect=create_engine("mysql+pymysql://root:1qaz@WSX@172.18.1.101/fisher?charset=utf8",echo=True)
 Session = sessionmaker(bind=db_connect)
 session=Session()
-def WarnPrices():
-    # session.execute('select * from Prices where DATEDIFF("2019-11-27",stopTime)<=0')
-    priwar=PricesWarn()
-    pricesList=session.query(Prices).filter(db.cast(Prices.stopTime, db.Date) <= db.cast(datetime.date.today(), db.Date)).all()
+def WarnPriceAtrr(pricesList):
+
+    prilist=[]
     for price in pricesList:
+        priwar = PricesWarn()
         setattr(priwar,"student_name",price.pri_stu.student_name)
         if price.pri_tea!=None:
             setattr(priwar,"teacher_name",price.pri_tea.teacher_name)
@@ -38,22 +38,26 @@ def WarnPrices():
         setattr(priwar,"stopTime",price.stopTime)
         setattr(priwar,"ClassHours",price.ClassHours)
         setattr(priwar,"prices",price.prices)
+        prilist.append(priwar)
+    return prilist
+def WarnPrices():
+    # session.execute('select * from Prices where DATEDIFF("2019-11-27",stopTime)<=0')
+
+    #cast函数可以将时间转换成日期
+    pricesList=session.query(Prices).order_by(Prices.stopTime).filter(db.cast(Prices.stopTime, db.Date) <= db.cast(datetime.date.today()-datetime.timedelta(days=7), db.Date)).all()
+    if pricesList==[]:
+        return
     result=session.query(PricesWarn).all()
     try:
         if len(result)!=0:
             session.execute("TRUNCATE table PricesWarn;")
-        session.add(priwar)
+        session.add_all(WarnPriceAtrr(pricesList))
         session.commit()
     except Exception as e:
         print(e)
 
 def accept_pattern(content,pattern_str):
     return re.match(pattern_str,content).group(1)
-
-
-#db = SQLAlchemy()
-
-
 
 
 def create_app(configuration):
@@ -84,5 +88,5 @@ app = create_app(os.getenv('ENVIRONMENT'))
 Schuduler = APScheduler()  #定时任务
 Schuduler.init_app(app)
 
-api = Api(app=app, prefix="/api/v1")
+# api = Api(app=app, prefix="/api/v1")
 from app import views
